WidgetsDirectory доработан, чтобы отличать разные справочники. По недосмотру все справочники действовали как один. WidgetsDirectory: добавлена проверка на длины строк. Сами длины строк вынесены в константы.

+++ plugins/widgets/lib/directorycontroller.php
@@ -4,6 +4,8 @@ namespace Difra\Plugins\Widgets;
 
 abstract class DirectoryController extends \Difra\Controller {
 
+	const directory = null;
+
 	abstract public function action( $value );
 
 	public function indexAjaxAction() {
@@ -15,9 +17,19 @@ abstract class DirectoryController extends \Difra\Controller {
 	public function addAjaxAction( \Difra\Param\AjaxString $search ) {
 
 		$this->subInit();
+		/** @var \Difra\Plugins\Widgets\Objects\Directory $class */
 		$class = \Difra\Unify\Storage::getClass( 'WidgetsDirectory' );
-		if( !$object = $class::getByField( 'name', $search ) ) {
+		if( strlen( $search ) > $class::DIRECTORY_LENGTH ) {
+			\Difra\Ajax::getInstance()->notify(
+				\Difra\Locales::getInstance()->getXPath( 'widgets/directory/value-too-long' )
+			);
+		}
+		$searchObj = new \Difra\Unify\Search( 'WidgetsDirectory' );
+		$searchObj->addConditions( array( 'directory' => static::directory, 'name' => $search ) );
+		$res = $searchObj->doQuery();
+		if( empty( $res ) ) {
 			$object = $class::create();
+			$object->directory = static::directory;
 			$object->name = (string)$search;
 		}
 		\Difra\Ajax::getInstance()->close();
@@ -30,6 +42,9 @@ abstract class DirectoryController extends \Difra\Controller {
 		try {
 			$class = \Difra\Unify\Storage::getClass( 'WidgetsDirectory' );
 			$object = $class::get( (string)$id );
+			if( $object->directory != static::directory ) {
+				throw new \Difra\Exception( 'This item does not exist in this directory.' );
+			}
 			\Difra\Ajax::getInstance()->close();
 			$this->action( $object->name );
 		} catch( \Difra\Exception $ex ) {
@@ -43,6 +58,9 @@ abstract class DirectoryController extends \Difra\Controller {
 		try {
 			$class = \Difra\Unify\Storage::getClass( 'WidgetsDirectory' );
 			$object = $class::get( (string)$id );
+			if( $object->directory != static::directory ) {
+				throw new \Difra\Exception( 'This item does not exist in this directory.' );
+			}
 			$object->delete();
 		} catch( \Difra\Exception $ex ) {
 		}
@@ -54,14 +72,18 @@ abstract class DirectoryController extends \Difra\Controller {
 		$xml = new \DOMDocument();
 		$node = $xml->appendChild( $xml->createElement( 'DirectoryWindow' ) );
 		$search = new \Difra\Unify\Search( 'WidgetsDirectory' );
+		$search->addCondition( 'directory', static::directory );
 		$search->getListXML( $node );
 		return \Difra\View::render( $xml, 'widget_directory', true );
 	}
 
 	private function subInit() {
 
-		if( !defined( 'static::directory' ) ) {
+		if( !static::directory ) {
 			throw new \Difra\Exception( 'DirectoryController extended class should have \'directory\' constant with directory name.' );
+		}
+		if( strlen( static::directory ) > ( $len = \Difra\Plugins\Widgets\Objects\Directory::DIRECTORY_LENGTH ) ) {
+			throw new \Difra\Exception( 'WidgetsDirectory directory name is too long. ' . $len . ' bytes is the limit.' );
 		}
 	}
 }
